Add Git Tasks.
Add Tasks:
- Git: Begin interactive rebase with autostash
  - Allows selecting ‘--rebase-merges’.
- Git: Continue rebase
- Git: Abort rebase
- Git: Commit Staged (Amend)
  - Allows editing the commit message as opposed to the respective
    command of VS Code's SCM.
- Git: Commit using mtime
  - Uses script common/git-commit-using-mtime so the "Authored" date
    will be the date when the file to be committed has been modified,
    as opposed to VS Code's commit command which used the current date.
    If multiple files are to be committed, the latest mtime will be used.

All Tasks run the respective command in the dir of the file opened in
the active editor, not in the workspace root (the default).
This is crucial for workspaces like "scripts".

Furthermore, the respective command is run via
common/exec-in-git-repo-root-dir.
This script causes the command to succeed even if the currently opened
file resides in a (sub)+dir of the repo.